Migrated Player_Base C++ class
Player_Base class is now the parent class for both the pc and vr turtle blueprints. it contains all the common functionality as well new audio logic and improved performance. Other classes can now refer to the base class if getting a variable that belongs to the base class rather than doing a enum and switch